Updated model (sobject, particpant, layout, design) and & regenerated
As sobjects can occur in many places with the same parameters, the pose was moved to the attraction participant. The layout was simplified to only have attraction trees. On the design, element trees have been added in order to allow elements to again seperate from their pose. Further element trees allow to store part-whole relationships between elements, allowing elements of several lods to be present in the same design and having semantics. Both should help the UI.