Replaced ElementTree with ElementOccurance
The recursion of elements do not happen on design level but possibly on element level. A design therefore shouldn't reflect the part to whole representation (as they did with element trees.). The only way to (currently) keep the information is by adding concepts to the representations from which the relationships can be constructed.