Added fork of speckle rhino converter
The speckle converter is tied to the RhinoDoc which is a singleton with RhinoDoc.ActiveDoc which gives crap when running running inside Compute. As a consequence the serilalization would work in a normal gh document but not in Compute. I had to remove all doc (only purpose was to provide units) related content and stripped it down to a pure converter.