Enforced naming convention in Python implementations
The file name of an implementation of a service can't have be the service name because it would wrongly import the base implementation.
Therefore by convention: semioSERVICENAME.py for the implementation of the service. E.g. semiomanager.py for the manager service
For extensions: semioextensionPLATFORMNAME e.g semioextensiongrasshopper.py for a Grasshopper extension.