🛣️Big meta model change (separated geometry and semantics)
- Removed unification of Properties (scalar, geometry, file, container, ...) because lots of unintended side effects:
Ports can have complex Properties which are not intended.
Big problem: Geometry
Should geometry be present as scalars? What would be the difference/advantage compared to a FileProperty?
- Conclusion:
(Semantic) Properties for Types are now Qualities. These should be accessible for transformations.
GeometryProperties are now Representations (again) and are not (easily) accessible inside transformations.